requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
torch==2.8.0+cpu
torchvision==0.23.0+cpu
python-multipart==0.0.20
Pillow==11.3.0
